+++ src/Discord.Net.ComponentDesigner.Generator/Parsing/Parser2/BlendedNode.cs
@@ -1,0 +1,7 @@
+﻿namespace Discord.ComponentDesignerGenerator.Parser;
+
+public readonly record struct BlendedNode(
+    CXNode? Node,
+    CXToken? Token,
+    CXBlender Blender
+);
