+++ src/Discord.Net.ComponentDesigner.Generator/Parsing/Parser2/CXBlender.cs
@@ -1,144 +1,212 @@
 ﻿using Microsoft.CodeAnalysis.Text;
 using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Diagnostics;
 
 namespace Discord.ComponentDesignerGenerator.Parser;
 
-public sealed class CXBlender
+public readonly struct CXBlender
 {
-    public Queue<TextChange> Changes { get; }
+    private readonly CXLexer _lexer;
+    private readonly ImmutableStack<TextChangeRange> _changes;
 
-    private int CurrentSourcePosition => Lexer.Reader.Position;
-    public CXLexer Lexer => Document.Parser.Lexer;
+    private readonly int _newPosition;
+    private readonly int _changeDelta;
 
-    public CXDoc Document { get; }
+    private readonly CXCursor _cursor;
 
-    private readonly List<CXToken> _tokens;
 
-    private int _docTokenIndex;
-
-    private int _changeDelta;
-    
     public CXBlender(
-        CXDoc document
+        CXLexer lexer,
+        CXDoc document,
+        IEnumerable<TextChangeRange> changes,
+        CXCursor? cursor = null
     )
     {
-        _tokens = [];
-        Document = document;
-        Changes = [];
+        _lexer = lexer;
+
+        _newPosition = _lexer.Reader.Source.SourceSpan.Start;
+
+        _changes = [..changes];
+
+        _cursor = cursor ?? CXCursor.FromRoot(document).MoveToFirstChild();
     }
 
-    public void Reset()
+    private CXBlender(
+        CXLexer lexer,
+        CXCursor cursor,
+        ImmutableStack<TextChangeRange> changes,
+        int newPosition,
+        int changeDelta
+    )
     {
-        _tokens.Clear();
-        _docTokenIndex = 0;
-        _changeDelta = 0;
-        Changes.Clear();
+        _lexer = lexer;
+        _cursor = cursor;
+        _changes = changes;
+        _newPosition = newPosition;
+        _changeDelta = changeDelta;
     }
 
-    public CXToken GetToken(int index)
+    public static TextChangeRange GetAffectedRange(CXDoc doc, TextChangeRange range)
     {
-        while (_tokens.Count <= index)
+        return range;
+
+        // // clamp the range to the end of the doc
+        // var start = Math.Max(Math.Min(range.Span.Start, doc.FullSpan.End - 1), 0);
+        //
+        // if (!doc.TryFindToken(start, out var token)) return range;
+        //
+        // start = Math.Max(0, token.Span.Start - 1);
+        //
+        // var span = TextSpan.FromBounds(start, range.Span.End);
+        // var length = range.NewLength + (range.Span.Start - start);
+        // return new(span, length);
+    }
+
+    public BlendedNode ReadNode() => ReadNodeOrToken(asToken: false);
+    public BlendedNode ReadToken() => ReadNodeOrToken(asToken: true);
+
+    private BlendedNode ReadNodeOrToken(bool asToken)
+        => new Reader(this).ReadNodeOrToken(asToken);
+
+    public struct Reader
+    {
+        private CXCursor _oldCursor;
+        private ImmutableStack<TextChangeRange> _changes;
+        private int _newPosition;
+        private int _changeDelta;
+
+        private readonly CXLexer _lexer;
+
+        public Reader(CXBlender blender)
         {
-            var token = NextToken();
-
-            if (token.Kind is CXTokenKind.EOF) return token;
+            _lexer = blender._lexer;
+            _oldCursor = blender._cursor;
+            _changes = blender._changes;
+            _newPosition = blender._newPosition;
+            _changeDelta = blender._changeDelta;
         }
 
-        return _tokens[index];
-    }
-
-    public CXToken NextToken()
-    {
-        SkipPastChanges();
-
-        while (true)
+        public BlendedNode ReadNodeOrToken(bool asToken)
         {
-            while (_changeDelta < 0 && _docTokenIndex < Document.Tokens.Count)
+            while (true)
             {
-                var oldToken = Document.Tokens[_docTokenIndex++];
-                _changeDelta += oldToken.AbsoluteWidth;
+                if (_oldCursor.IsDone) return ReadNewToken();
+
+                if (_changeDelta < 0) SkipOldToken();
+                else if (_changeDelta > 0) return ReadNewToken();
+                else
+                {
+                    if (TryTakeOldNodeOrToken(asToken, out var blendedNode)) return blendedNode;
+
+                    if (_oldCursor.Current.Node is not null)
+                        _oldCursor = _oldCursor.MoveToFirstChild();
+                    else
+                        SkipOldToken();
+                }
             }
-
-            if (_changeDelta > 0)
-                return LexNewToken();
-
-            if (TryReuseToken(out var token)) return token;
-
-            if (Document.Tokens.Count <= _docTokenIndex) return LexNewToken();
-
-            _changeDelta += Document.Tokens[_docTokenIndex++].AbsoluteWidth;
         }
 
-        bool TryReuseToken(out CXToken token)
+        private void SkipOldToken()
         {
-            if (_docTokenIndex >= Document.Tokens.Count)
+            _oldCursor = _oldCursor.MoveToFirstToken();
+
+            var current = _oldCursor.Current;
+
+            _changeDelta += current.FullSpan.Length;
+
+            _oldCursor = CXCursor.MoveToNextSibling(_oldCursor);
+
+            SkipPastChanges();
+        }
+
+        private void SkipPastChanges()
+        {
+            var oldPosition = _oldCursor.Current.FullSpan.Start;
+
+            while (!_changes.IsEmpty && oldPosition >= _changes.Peek().Span.End)
             {
-                token = default;
+                var change = _changes.Peek();
+
+                _changes = _changes.Pop();
+                _changeDelta += change.NewLength - change.Span.Length;
+            }
+        }
+
+        private BlendedNode ReadNewToken()
+        {
+            var token = LexNewToken();
+
+            _newPosition += token.FullSpan.Length;
+            _changeDelta -= token.FullSpan.Length;
+
+            SkipPastChanges();
+
+            return CreateBlendedNode(token: token);
+        }
+
+        private CXToken LexNewToken()
+        {
+            _lexer.Reader.Position = _newPosition;
+            return _lexer.Next();
+        }
+
+        private bool TryTakeOldNodeOrToken(
+            bool asToken,
+            out BlendedNode blendedNode
+        )
+        {
+            if (asToken) _oldCursor = _oldCursor.MoveToFirstToken();
+
+            var current = _oldCursor.Current;
+
+            if (!CanReuse(current))
+            {
+                blendedNode = default;
                 return false;
             }
 
-            token = Document.Tokens[_docTokenIndex];
+            _newPosition += current.FullSpan.Length;
+            _oldCursor = CXCursor.MoveToNextSibling(_oldCursor);
 
-            if (!CanReuse(token)) return false;
-
-            _docTokenIndex++;
-            Lexer.Reader.Advance(token.AbsoluteWidth);
-
-            _tokens.Add(token);
+            blendedNode = CreateBlendedNode(
+                node: current.Node,
+                token: current.Token
+            );
             return true;
         }
-    }
 
-    private CXToken LexNewToken()
-    {
-        var token = Lexer.Next();
-
-        _tokens.Add(token);
-        _changeDelta += token.AbsoluteWidth;
-
-        return token;
-    }
-
-    private void SkipPastChanges()
-    {
-        while (Changes.Count > 0)
+        private bool CanReuse(NodeOrToken value)
         {
-            var change = Changes.Peek();
-            var newLength = change.NewText?.Length ?? 0;
+            if (!value.HasValue) return false;
 
-            if (change.Span.Start + newLength > CurrentSourcePosition)
-                break;
+            if (value.FullSpan.IsEmpty) return false;
 
-            Changes.Dequeue();
+            if (IntersectsNextChange(value.FullSpan)) return false;
 
-            _changeDelta += newLength - change.Span.Length;
+            // TODO: more riggerous checking; no error nodes/tokens, etc
 
-            // update the cursor to the new change
-            while (_docTokenIndex < Document.Tokens.Count)
-            {
-                var token = Document.Tokens[_docTokenIndex];
-
-                if (token.AbsoluteStart >= change.Span.Start)
-                    break;
-
-                _docTokenIndex++;
-            }
+            return true;
         }
-    }
 
-    private bool CanReuse(CXToken token)
-    {
-        if (token.AbsoluteWidth is 0) return false;
+        private bool IntersectsNextChange(TextSpan span)
+            => !_changes.IsEmpty && span.IntersectsWith(_changes.Peek().Span);
 
-        if (IntersectsNextChange(token.Span)) return false;
+        private BlendedNode CreateBlendedNode(CXNode? node = null, CXToken? token = null)
+        {
+            Debug.Assert(node is not null || token.HasValue);
 
-        return true;
-    }
-
-    private bool IntersectsNextChange(TextSpan span)
-    {
-        if (Changes.Count is 0) return false;
-
-        return span.IntersectsWith(Changes.Peek().Span);
+            return new(
+                node,
+                token,
+                new CXBlender(
+                    _lexer,
+                    _oldCursor,
+                    _changes,
+                    _newPosition,
+                    _changeDelta
+                )
+            );
+        }
     }
 }
