+++ src/Discord.Net.ComponentDesigner.Generator/Parsing/Parser2/IncrementalParseContext.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.CodeAnalysis.Text;
+using System.Collections.Generic;
+
+namespace Discord.ComponentDesignerGenerator.Parser;
+
+public readonly record struct IncrementalParseContext(
+    IReadOnlyList<TextChange> Changes,
+    TextChangeRange AffectedRange
+);
